reorg and standardization, minimizing code in react files
