+++ templates/react/src/example/Example.tsx
@@ -1,63 +1,13 @@
 import React, { useCallback, useEffect, useState } from 'react'
 import './Example.css'
-import { ethers } from 'ethers'
-import { AlphaRouter, ChainId, SwapType } from '@uniswap/smart-order-router'
-import { TradeType, CurrencyAmount, Percent } from '@uniswap/sdk-core'
 import { Environment, CurrentConfig } from '../config'
 import { getCurrencyBalance } from '../libs/wallet'
 import {
   connectBrowserExtensionWallet,
-  getMainnetProvider,
   getProvider,
   getWalletAddress,
-  sendTransaction,
   TransactionState,
 } from '../libs/providers'
-import {
-  V3_SWAP_ROUTER_ADDRESS,
-  MAX_FEE_PER_GAS,
-  MAX_PRIORITY_FEE_PER_GAS,
-} from '../libs/constants'
-
-async function route(): Promise<TransactionState> {
-  const router = new AlphaRouter({
-    chainId: ChainId.MAINNET,
-    provider: getMainnetProvider(),
-  })
-  const address = getWalletAddress()
-
-  if (!address) {
-    return TransactionState.Failed
-  }
-
-  const route = await router.route(
-    CurrencyAmount.fromRawAmount(
-      CurrentConfig.currencies.in,
-      ethers.utils
-        .parseEther(CurrentConfig.currencies.amountIn.toString())
-        .toString()
-    ),
-    CurrentConfig.currencies.out,
-    TradeType.EXACT_INPUT,
-    {
-      recipient: CurrentConfig.wallet.address,
-      slippageTolerance: new Percent(5, 100),
-      deadline: Math.floor(Date.now() / 1000 + 1800),
-      type: SwapType.SWAP_ROUTER_02,
-    }
-  )
-
-  const res = await sendTransaction({
-    data: route?.methodParameters?.calldata,
-    to: V3_SWAP_ROUTER_ADDRESS,
-    value: route?.methodParameters?.value,
-    from: address,
-    maxFeePerGas: MAX_FEE_PER_GAS,
-    maxPriorityFeePerGas: MAX_PRIORITY_FEE_PER_GAS,
-  })
-
-  return res
-}
 
 const useOnBlockUpdated = (callback: (blockNumber: number) => void) => {
   useEffect(() => {
@@ -106,9 +56,9 @@ const Example = () => {
     }
   }, [refreshBalances])
 
-  const onTrade = useCallback(async () => {
-    setTxState(TransactionState.Sending)
-    setTxState(await route())
+  const onAction = useCallback(async () => {
+    // TODO do your action here
+    setTxState(TransactionState.Sent)
   }, [])
 
   return (
@@ -132,7 +82,7 @@ const Example = () => {
       <h3>{`Token In (${CurrentConfig.currencies.in.symbol}) Balance: ${tokenInBalance}`}</h3>
       <h3>{`Token Out (${CurrentConfig.currencies.out.symbol}) Balance: ${tokenOutBalance}`}</h3>
       <button
-        onClick={onTrade}
+        onClick={onAction}
         disabled={
           txState === TransactionState.Sending ||
           getProvider() === null ||
