+++ templates/react/src/libs/constants.ts
@@ -4,12 +4,26 @@ import { Ether, SupportedChainId, Token } from '@uniswap/sdk-core'
 
 // Addresses
 
+export const POOL_FACTORY_CONTRACT_ADDRESS =
+  '0x1F98431c8aD98523631AE4a59f267346ea31F984'
+export const QUOTER_CONTRACT_ADDRESS =
+  '0xb27308f9F90D607463bb33eA1BeBb41C27CE5AB6'
 export const V3_SWAP_ROUTER_ADDRESS =
   '0x68b3465833fb72A70ecDF485E0e4C7bD8665Fc45'
+export const WETH_CONTRACT_ADDRESS =
+  '0xc02aaa39b223fe8d0a0e5c4f27ead9083c756cc2'
 
 // Currencies and Tokens
 
 export const ETH = Ether.onChain(SupportedChainId.MAINNET)
+
+export const WETH_TOKEN = new Token(
+  SupportedChainId.MAINNET,
+  '0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2',
+  18,
+  'WETH',
+  'Wrapped Ether'
+)
 
 export const USDC_TOKEN = new Token(
   SupportedChainId.MAINNET,
@@ -21,7 +35,7 @@ export const USDC_TOKEN = new Token(
 
 // ABI's
 
-export const ERC_20_ABI = [
+export const ERC20_ABI = [
   // Read-Only Functions
   'function balanceOf(address owner) view returns (uint256)',
   'function decimals() view returns (uint8)',
@@ -32,6 +46,14 @@ export const ERC_20_ABI = [
 
   // Events
   'event Transfer(address indexed from, address indexed to, uint amount)',
+]
+
+export const WETH_ABI = [
+  // Wrap ETH
+  'function deposit() payable',
+
+  // Unwrap ETH
+  'function withdraw(uint wad) public',
 ]
 
 // Transactions
