+++ templates/react/src/libs/wallet.ts
@@ -1,10 +1,18 @@
 // This file contains code to easily connect to and get information from a wallet on chain
 
 import { Currency } from '@uniswap/sdk-core'
-import { ethers } from 'ethers'
+import { BigNumber, ethers } from 'ethers'
 import { providers } from 'ethers'
-import { ERC_20_ABI } from './constants'
+import JSBI from 'jsbi'
+import {
+  ERC20_ABI,
+  MAX_FEE_PER_GAS,
+  MAX_PRIORITY_FEE_PER_GAS,
+  WETH_ABI,
+  WETH_CONTRACT_ADDRESS,
+} from './constants'
 import { toReadableAmount } from './conversion'
+import { getProvider, getWalletAddress, sendTransaction } from './providers'
 
 export async function getCurrencyBalance(
   provider: providers.Provider,
@@ -19,7 +27,7 @@ export async function getCurrencyBalance(
   // Get currency otherwise
   const currencyContract = new ethers.Contract(
     currency.address,
-    ERC_20_ABI,
+    ERC20_ABI,
     provider
   )
   const balance: number = await currencyContract.balanceOf(address)
@@ -27,4 +35,61 @@ export async function getCurrencyBalance(
 
   // Format with proper units (approximate)
   return toReadableAmount(balance, decimals).toString()
+}
+
+// wraps ETH (rounding up to the nearest ETH for decimal places)
+export async function wrapETH(eth: number) {
+  const provider = getProvider()
+  const address = getWalletAddress()
+  if (!provider || !address) {
+    throw new Error('Cannot wrap ETH without a provider and wallet address')
+  }
+
+  const wethContract = new ethers.Contract(
+    WETH_CONTRACT_ADDRESS,
+    WETH_ABI,
+    provider
+  )
+
+  const transaction = {
+    data: wethContract.interface.encodeFunctionData('deposit'),
+    value: BigNumber.from(Math.ceil(eth))
+      .mul(JSBI.exponentiate(JSBI.BigInt(10), JSBI.BigInt(18)).toString())
+      .toString(),
+    from: address,
+    to: WETH_CONTRACT_ADDRESS,
+    maxFeePerGas: MAX_FEE_PER_GAS,
+    maxPriorityFeePerGas: MAX_PRIORITY_FEE_PER_GAS,
+  }
+
+  await sendTransaction(transaction)
+}
+
+// unwraps ETH (rounding up to the nearest ETH for decimal places)
+export async function unwrapETH(eth: number) {
+  const provider = getProvider()
+  const address = getWalletAddress()
+  if (!provider || !address) {
+    throw new Error('Cannot unwrap ETH without a provider and wallet address')
+  }
+
+  const wethContract = new ethers.Contract(
+    WETH_CONTRACT_ADDRESS,
+    WETH_ABI,
+    provider
+  )
+
+  const transaction = {
+    data: wethContract.interface.encodeFunctionData('withdraw', [
+      BigNumber.from(Math.ceil(eth))
+        .mul(JSBI.exponentiate(JSBI.BigInt(10), JSBI.BigInt(18)).toString())
+        .toString(),
+    ]),
+    from: address,
+    to: WETH_CONTRACT_ADDRESS,
+    maxFeePerGas: MAX_FEE_PER_GAS,
+    maxPriorityFeePerGas: MAX_PRIORITY_FEE_PER_GAS,
+  }
+
+  await sendTransaction(transaction)
 }
