+++ v3-sdk/collecting-fees/src/libs/liquidity.ts
@@ -14,12 +14,58 @@ import {
   nearestUsableTick,
   MintOptions,
   NonfungiblePositionManager,
+  CollectOptions,
 } from '@uniswap/v3-sdk'
 import { CurrentConfig } from '../config'
 import { getPoolInfo } from './pool'
 import { getProvider, getWalletAddress } from './providers'
 import { Percent, CurrencyAmount, Token } from '@uniswap/sdk-core'
 import { fromReadableAmount } from './conversion'
+
+export async function collectFees(
+  positionId: number
+): Promise<TransactionState> {
+  const address = getWalletAddress()
+  const provider = getProvider()
+  if (!address || !provider) {
+    return TransactionState.Failed
+  }
+
+  const collectOptions: CollectOptions = {
+    tokenId: positionId,
+    expectedCurrencyOwed0: CurrencyAmount.fromRawAmount(
+      CurrentConfig.tokens.token0,
+      fromReadableAmount(
+        CurrentConfig.tokens.token0AmountToCollect,
+        CurrentConfig.tokens.token0.decimals
+      )
+    ),
+    expectedCurrencyOwed1: CurrencyAmount.fromRawAmount(
+      CurrentConfig.tokens.token1,
+      fromReadableAmount(
+        CurrentConfig.tokens.token1AmountToCollect,
+        CurrentConfig.tokens.token1.decimals
+      )
+    ),
+    recipient: address,
+  }
+
+  // get calldata for minting a position
+  const { calldata, value } =
+    NonfungiblePositionManager.collectCallParameters(collectOptions)
+
+  // build transaction
+  const transaction = {
+    data: calldata,
+    to: NONFUNGIBLE_POSITION_MANAGER_CONTRACT_ADDRESS,
+    value: value,
+    from: address,
+    maxFeePerGas: MAX_FEE_PER_GAS,
+    maxPriorityFeePerGas: MAX_PRIORITY_FEE_PER_GAS,
+  }
+
+  return sendTransaction(transaction)
+}
 
 export async function getPositionIds(
   provider: ethers.providers.Provider,
@@ -65,7 +111,7 @@ export async function getTokenTransferApprovals(
       TOKEN_AMOUNT_TO_APPROVE_FOR_TRANSFER
     )
 
-    await sendTransaction({
+    return sendTransaction({
       ...transaction,
       from: fromAddress,
     })
@@ -73,19 +119,17 @@ export async function getTokenTransferApprovals(
     console.error(e)
     return TransactionState.Failed
   }
-
-  return TransactionState.Sent
 }
 
-export const constructPosition = async (
+export async function constructPosition(
   token0Amount: CurrencyAmount<Token>,
   token1Amount: CurrencyAmount<Token>
-): Promise<Position> => {
+): Promise<Position> {
   // get pool info
   const poolInfo = await getPoolInfo()
 
   // construct pool instance
-  const USDC_DAI_POOL = new Pool(
+  const configuredPool = new Pool(
     token0Amount.currency,
     token1Amount.currency,
     poolInfo.fee,
@@ -96,7 +140,7 @@ export const constructPosition = async (
 
   // create position using the maximum liquidity from input amounts
   return Position.fromAmounts({
-    pool: USDC_DAI_POOL,
+    pool: configuredPool,
     tickLower:
       nearestUsableTick(poolInfo.tick, poolInfo.tickSpacing) -
       poolInfo.tickSpacing * 2,
@@ -175,6 +219,5 @@ export async function mintPosition(): Promise<TransactionState> {
     maxPriorityFeePerGas: MAX_PRIORITY_FEE_PER_GAS,
   }
 
-  await sendTransaction(transaction)
-  return TransactionState.Sent
+  return sendTransaction(transaction)
 }
