+++ v3-sdk/minting-position/src/libs/positions.ts
@@ -1,7 +1,113 @@
+import { Percent } from '@uniswap/sdk-core'
+import {
+  nearestUsableTick,
+  NonfungiblePositionManager,
+  Pool,
+  Position,
+} from '@uniswap/v3-sdk'
 import { ethers } from 'ethers'
-import { ERC20_ABI, NONFUNGIBLE_POSITION_MANAGER_ABI } from './constants'
+import { CurrentConfig } from '../config'
+import {
+  ERC20_ABI,
+  MAX_FEE_PER_GAS,
+  MAX_PRIORITY_FEE_PER_GAS,
+  NONFUNGIBLE_POSITION_MANAGER_ABI,
+  NONFUNGIBLE_POSITION_MANAGER_CONTRACT_ADDRESS,
+} from './constants'
 import { AMOUNT_TO_APPROVE } from './constants'
-import { sendTransaction, TransactionState } from './providers'
+import { fromReadableAmount } from './conversion'
+import { getPoolInfo } from './pool'
+import {
+  getProvider,
+  getWalletAddress,
+  sendTransaction,
+  TransactionState,
+} from './providers'
+
+export async function mintPosition(): Promise<TransactionState> {
+  const address = getWalletAddress()
+  const provider = getProvider()
+  if (!address || !provider) {
+    return TransactionState.Failed
+  }
+
+  // Give approval to the contract to transfer tokens
+  const tokenInApproval = await getTokenTransferApprovals(
+    provider,
+    CurrentConfig.tokens.token0.address,
+    address,
+    NONFUNGIBLE_POSITION_MANAGER_CONTRACT_ADDRESS
+  )
+  const tokenOutApproval = await getTokenTransferApprovals(
+    provider,
+    CurrentConfig.tokens.token1.address,
+    address,
+    NONFUNGIBLE_POSITION_MANAGER_CONTRACT_ADDRESS
+  )
+
+  // Fail if transfer approvals do not go through
+  if (
+    tokenInApproval !== TransactionState.Sent ||
+    tokenOutApproval !== TransactionState.Sent
+  ) {
+    return TransactionState.Failed
+  }
+
+  // get pool data
+  const poolInfo = await getPoolInfo()
+
+  // create Pool abstraction
+  const configuredPool = new Pool(
+    CurrentConfig.tokens.token0,
+    CurrentConfig.tokens.token1,
+    poolInfo.fee,
+    poolInfo.sqrtPriceX96.toString(),
+    poolInfo.liquidity.toString(),
+    poolInfo.tick
+  )
+
+  // create position using the maximum liquidity from input amounts
+  const position = Position.fromAmounts({
+    pool: configuredPool,
+    tickLower:
+      nearestUsableTick(poolInfo.tick, poolInfo.tickSpacing) -
+      poolInfo.tickSpacing * 2,
+    tickUpper:
+      nearestUsableTick(poolInfo.tick, poolInfo.tickSpacing) +
+      poolInfo.tickSpacing * 2,
+    amount0: fromReadableAmount(
+      CurrentConfig.tokens.token0Amount,
+      CurrentConfig.tokens.token0.decimals
+    ),
+    amount1: fromReadableAmount(
+      CurrentConfig.tokens.token1Amount,
+      CurrentConfig.tokens.token1.decimals
+    ),
+    useFullPrecision: true,
+  })
+
+  // get calldata for minting a position
+  const { calldata, value } = NonfungiblePositionManager.addCallParameters(
+    position,
+    {
+      recipient: address,
+      deadline: Math.floor(Date.now() / 1000) + 60 * 20,
+      slippageTolerance: new Percent(50, 10_000),
+    }
+  )
+
+  // build transaction
+  const transaction = {
+    data: calldata,
+    to: NONFUNGIBLE_POSITION_MANAGER_CONTRACT_ADDRESS,
+    value: value,
+    from: address,
+    maxFeePerGas: MAX_FEE_PER_GAS,
+    maxPriorityFeePerGas: MAX_PRIORITY_FEE_PER_GAS,
+  }
+
+  return sendTransaction(transaction)
+}
 
 export async function getPositionIds(
   provider: ethers.providers.Provider,
@@ -47,7 +153,7 @@ export async function getTokenTransferApprovals(
       AMOUNT_TO_APPROVE
     )
 
-    await sendTransaction({
+    return sendTransaction({
       ...transaction,
       from: fromAddress,
     })
@@ -55,6 +161,4 @@ export async function getTokenTransferApprovals(
     console.error(e)
     return TransactionState.Failed
   }
-
-  return TransactionState.Sent
 }
