+++ v3-sdk/quoting/src/example/Example.tsx
@@ -1,68 +1,7 @@
 import React, { useState, useCallback } from 'react'
 import './Example.css'
-import { ethers } from 'ethers'
 import { CurrentConfig } from '../config'
-import { computePoolAddress } from '@uniswap/v3-sdk'
-import Quoter from '@uniswap/v3-periphery/artifacts/contracts/lens/Quoter.sol/Quoter.json'
-import IUniswapV3PoolABI from '@uniswap/v3-core/artifacts/contracts/interfaces/IUniswapV3Pool.sol/IUniswapV3Pool.json'
-import {
-  POOL_FACTORY_CONTRACT_ADDRESS,
-  QUOTER_CONTRACT_ADDRESS,
-} from '../libs/constants'
-import { getProvider } from '../libs/providers'
-import { toReadableAmount, fromReadableAmount } from '../libs/conversion'
-
-const getPoolConstants = async (): Promise<{
-  token0: string
-  token1: string
-  fee: number
-}> => {
-  const currentPoolAddress = computePoolAddress({
-    factoryAddress: POOL_FACTORY_CONTRACT_ADDRESS,
-    tokenA: CurrentConfig.tokens.in,
-    tokenB: CurrentConfig.tokens.out,
-    fee: CurrentConfig.tokens.fee,
-  })
-
-  const poolContract = new ethers.Contract(
-    currentPoolAddress,
-    IUniswapV3PoolABI.abi,
-    getProvider()
-  )
-  const [token0, token1, fee] = await Promise.all([
-    poolContract.token0(),
-    poolContract.token1(),
-    poolContract.fee(),
-  ])
-
-  return {
-    token0,
-    token1,
-    fee,
-  }
-}
-
-const quote = async (): Promise<string> => {
-  const quoterContract = new ethers.Contract(
-    QUOTER_CONTRACT_ADDRESS,
-    Quoter.abi,
-    getProvider()
-  )
-  const poolConstants = await getPoolConstants()
-
-  const quotedAmountOut = await quoterContract.callStatic.quoteExactInputSingle(
-    poolConstants.token0,
-    poolConstants.token1,
-    poolConstants.fee,
-    fromReadableAmount(
-      CurrentConfig.tokens.amountIn,
-      CurrentConfig.tokens.in.decimals
-    ),
-    0
-  )
-
-  return toReadableAmount(quotedAmountOut, CurrentConfig.tokens.out.decimals)
-}
+import { quote } from '../libs/quote'
 
 const Example = () => {
   const [outputAmount, setOutputAmount] = useState<string>()
